Optimize io for pagerank
Currently writing data is the slowest part. This is mostly due to the
fact that `dec(i)` is allocating an array to do the conversion. In
general our IO is quite bad at doing operations without allocation.

MMap'ing the input array is a huge speed boost as well as taking care
that parsing is not allocating.