Shuffle the datafiles after writing.
We need to make sure that the file we are reading in is not cached and
in a distributed settings the simplest thing is to make sure that each
processor doesn't read in its own file. Reads on a single node might
still be hitting cache so, we should investigate a broader solution for
this.